Consolidate some flyout code into flyout_base.
